Update README.md
Combination of multi-kingdom features improved the predictability in CRC prescreen and bacterial-fungal interaction contributed to the tumorigenesis of CRC

This repository contains the codes used for the following paper:


Depends

R version 3.6.2
Python 3.7.1
The full list of R packages for the code is given in requirements_R.txt. These can be installed using:

pip install -r requirements_R.txt

The full list of Python packages for the code is given in requirements.txt. These can be installed using:

pip install -r requirements.txt